@@ -18,28 +18,28 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.0.1+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   provider: ^5.0.0
   image_picker: ^0.7.5
-  adaptive_dialog: ^0.10.0
+  adaptive_dialog: ^0.10.0+5
   font_awesome_flutter: ^9.0.0
   intl: ^0.17.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.3
   json_annotation: ^4.0.1
-  collection: ^1.14.13
-  auto_size_text: ^2.1.0
+  collection: ^1.15.0
+  auto_size_text: ^3.0.0-nullsafety.0
   precept_script:
     path: ../precept_script
 
 
 dev_dependencies:
-  test: ^1.16.5
   flutter_test:
     sdk: flutter
   mockito: ^5.0.7
