@@ -5,11 +5,11 @@ import 'package:precept_client/app/precept.dart';
 import 'package:precept_client/page/errorPage.dart';
 import 'package:precept_client/page/standardPage.dart';
 import 'package:precept_client/user/signInFactory.dart';
+import 'package:precept_script/common/exception.dart';
 import 'package:precept_script/common/log.dart';
 import 'package:precept_script/common/script/error.dart';
 import 'package:precept_script/inject/inject.dart';
 import 'package:precept_script/script/script.dart';
-import 'package:precept_script/signin/signIn.dart';
 
 /// Router for Precept.
 ///
@@ -31,21 +31,19 @@ import 'package:precept_script/signin/signIn.dart';
 //     );
 ///
 class PreceptRouter {
-  String _preSignInRoute;
+  String? _preSignInRoute;
 
   PreceptRouter();
 
   Route<dynamic> generateRoute(RouteSettings settings, BuildContext context) {
     logType(this.runtimeType)
         .d("Requested route is: ${settings.name} with arguments ${settings.arguments}.");
-    final Map<String, dynamic> pageArguments = settings.arguments ?? {};
+    final Map<String, dynamic> pageArguments = (settings.arguments==null) ? {} : settings.arguments as Map<String, dynamic>;
     if (settings.name == 'signIn') {
       _preSignInRoute = pageArguments['returnRoute'];
-      final PSignInOptions options = pageArguments['signInConfig'];
       final DataProvider dataProvider = pageArguments['dataProvider'];
       final SignInFactory pageFactory = inject<SignInFactory>();
       final pageWidget = pageFactory.signInPage(
-        options: options,
         pageArguments: pageArguments,
         dataProvider: dataProvider,
       );
@@ -59,8 +57,11 @@ class PreceptRouter {
     if (settings.name == 'emailSignIn') {
       final factory = inject<EmailSignInFactory>();
       final DataProvider dataProvider = pageArguments['dataProvider'];
+      if (_preSignInRoute==null){
+        throw PreceptException ('At this point, _preSignInRoute must be set');
+      }
       final pageWidget = factory.emailSignInPage(
-        successRoute: _preSignInRoute,
+        successRoute: _preSignInRoute!,
         failureRoute: 'tbd',
         pageArguments: pageArguments,
         dataProvider: dataProvider,
@@ -72,7 +73,7 @@ class PreceptRouter {
           ),
           builder: (_) => pageWidget);
     }
-    final PPage preceptPage = script.pages[settings.name];
+    final PPage? preceptPage = script.pages[settings.name];
     if (preceptPage == null) {
       return _routeNotRecognised(settings);
     }
@@ -129,7 +130,7 @@ class PreceptRouter {
 
 class PreceptRouterConfig {
   final bool preceptFirst;
-  final Function(RouteSettings settings) alternateRouter;
+  final Function(RouteSettings settings)? alternateRouter;
 
   const PreceptRouterConfig({this.preceptFirst = true, this.alternateRouter});
 }
