@@ -1,9 +1,12 @@
-import 'package:flutter/foundation.dart';
+import 'package:precept_client/app/precept.dart';
 import 'package:precept_client/binding/binding.dart';
+import 'package:precept_client/binding/listBinding.dart';
+import 'package:precept_client/binding/stringBinding.dart';
 import 'package:precept_client/common/exceptions.dart';
+import 'package:precept_script/common/exception.dart';
 import 'package:precept_script/data/converter/converter.dart';
-import 'package:precept_script/schema/schema.dart';
-import 'package:precept_script/script/script.dart';
+import 'package:precept_script/schema/field/field.dart';
+import 'package:precept_script/schema/field/string.dart';
 
 /// Connects a Field Widget to its data source, using a [ModelViewConverter] to provide conversion where needed.
 /// For structural simplicity, if the data type of the Field is the same as the data type in the data source,
@@ -12,31 +15,37 @@ import 'package:precept_script/script/script.dart';
 class ModelConnector<MODEL, VIEW> {
   final Binding<MODEL> binding;
   final ModelViewConverter<MODEL, VIEW> converter;
-  final PSchemaElement fieldSchema;
+  final PField fieldSchema;
 
   ModelConnector(
-      {@required this.binding, this.converter = const PassThroughConverter(), this.fieldSchema});
+      {required this.binding,
+      this.converter = const PassThroughConverter(),
+      required this.fieldSchema});
 
   /// It is generally better to keep the default settings of the Binding for consistency - using [readFromModel] is therefore
   /// preferred.  Use this method only if you specifically need to override the Binding settings
   VIEW readFromModelOverridingDefaults(
-      {MODEL defaultValue, bool allowNullReturn = false, bool createIfAbsent = true}) {
-    return converter.modelToView(binding.read(
+      {MODEL? defaultValue, bool allowNullReturn = false, bool createIfAbsent = true}) {
+    final model = binding.read(
         defaultValue: defaultValue,
         allowNullReturn: allowNullReturn,
-        createIfAbsent: createIfAbsent));
+        createIfAbsent: createIfAbsent);
+    if (model != null) return converter.modelToView(model);
+    throw PreceptException('Model cannot be null');
   }
 
   /// Flutter validation requires that a null is returned if there are no validation errors.
   /// If there are multiple errors, a concatenated String is returned.
-  String validate(VIEW inputData, PScript pScript) {
-    final validationMessages= converter.validate(inputData, fieldSchema, pScript);
+  String? validate(VIEW inputData) {
+    final validationMessages = converter.validate(inputData, fieldSchema, script);
     return (validationMessages.isEmpty) ? null : validationMessages.join(', ');
   }
 
   /// See also [readFromModelOverridingDefaults]
   VIEW readFromModel() {
-    return converter.modelToView(binding.read());
+    final model = binding.read();
+    if (model != null) return converter.modelToView(model);
+    throw PreceptException('Model cannot be null');
   }
 
   writeToModel(VIEW value) {
@@ -44,10 +53,20 @@ class ModelConnector<MODEL, VIEW> {
   }
 }
 
+
+/// This is a bit convoluted to get around null safety, but presents static data in place of
+/// dynamic data from a database, while keeping the same code structure
 class StaticConnector extends ModelConnector<String, String> {
   final String staticData;
 
-  StaticConnector(this.staticData);
+  StaticConnector(this.staticData)
+      : super(
+            binding: StringBinding.private(
+              property: 'not used',
+              firstLevelKey: 'x',
+              parent: ListBinding.private(firstLevelKey: 'x', property: 'not used'),
+            ),
+            fieldSchema: PString());
 
   @override
   String readFromModel() {
