@@ -13,10 +13,10 @@ const String defValue = "a default";
 const String updateValue = "an update";
 
 void main() {
-  Map<String, dynamic> data;
-  MutableDocument temporaryDocument;
-  RootBinding rootBinding;
-  ChangeListener changeListener;
+  late Map<String, dynamic> data;
+  late MutableDocument temporaryDocument;
+  late   RootBinding rootBinding;
+  late   ChangeListener changeListener;
 
   setUp(() {
     data = generateData();
@@ -26,14 +26,14 @@ void main() {
     rootBinding =
         RootBinding(data: data, editHost: temporaryDocument, id: "test");
     changeListener = ChangeListener();
-    temporaryDocument.addListener(changeListener.listenToChange());
+    temporaryDocument.addListener(changeListener.listenToChange);
   });
 
   group("StringBinding", () {
     group("Read", () {
       test("read with default settings, value exists", () {
         final String actual =
-        rootBinding.stringBinding(property: property).read();
+        rootBinding.stringBinding(property: property).read()!;
         final String expected = loadedValue;
         expect(actual, expected);
         expect(changeListener.changeCount, 1,
@@ -44,7 +44,7 @@ void main() {
 
       test("read with default settings, value does not exist", () {
         final String actual =
-        rootBinding.stringBinding(property: "no item").read();
+        rootBinding.stringBinding(property: "no item").read()!;
         final String expected = "";
         expect(actual, expected);
         expect(changeListener.changeCount, 1,
@@ -58,7 +58,7 @@ void main() {
         final String expected = loadedValue;
         final String actual = rootBinding
             .stringBinding(property: property)
-            .read(defaultValue: defaultValue);
+            .read(defaultValue: defaultValue)!;
         expect(actual, expected);
         expect(changeListener.changeCount, 1,
             reason: "creating StringBinding adds property");
@@ -96,7 +96,7 @@ void main() {
       test(
           "read with no default value, value does not exist, allowNull is false",
               () {
-            final String expected = null;
+            final String? expected = null;
             final actual = rootBinding
                 .stringBinding(property: "no item")
                 .read(allowNullReturn: true);
@@ -112,7 +112,7 @@ void main() {
         final itemBinding = rootBinding.stringBinding(property: property);
         String expected = updateValue;
         itemBinding.write(expected);
-        String result = itemBinding.read();
+        String? result = itemBinding.read();
         expect(result, expected);
         expect(changeListener.changeCount, 1,
             reason: "creating StringBinding adds property");
