@@ -2,6 +2,7 @@ import 'package:precept_client/binding/binding.dart';
 import 'package:precept_client/binding/connector.dart';
 import 'package:precept_client/binding/mapBinding.dart';
 import 'package:precept_client/data/dataBinding.dart';
+import 'package:precept_script/common/exception.dart';
 import 'package:precept_script/data/converter/converter.dart';
 import 'package:precept_script/part/part.dart';
 import 'package:precept_script/schema/field/field.dart';
@@ -11,18 +12,25 @@ import 'package:precept_script/schema/schema.dart';
 
 mixin ConnectorBuilder {
   ModelConnector buildConnector(
-      { DataBinding dataBinding, PPart config, Type viewDataType}) {
+      {required DataBinding dataBinding, required PPart config,required Type viewDataType}) {
     final ModelBinding parentBinding = dataBinding.binding;
     final PDocument schema = dataBinding.schema;
-    final PSchemaElement fieldSchema = schema.fields[config.property];
+    final String? propertyName=config.property;
+    if (propertyName==null){
+      throw PreceptException('propertyName is required');
+    }
+    final PSchemaElement? fieldSchema = schema.fields[config.property];
+    if (fieldSchema==null){
+      throw PreceptException('fieldSchema must be defined');
+    }
     final binding =
-        _binding(parentBinding: parentBinding, schema: fieldSchema, property: config.property);
+        _binding(parentBinding: parentBinding, schema: fieldSchema as PField, property: propertyName);
     final converter = _converter(schema: fieldSchema, viewDataType: viewDataType);
-    final connector = ModelConnector(binding: binding, converter: converter);
+    final connector = ModelConnector(binding: binding, converter: converter,fieldSchema: fieldSchema);
     return connector;
   }
 
-  Binding _binding({ModelBinding parentBinding, PField schema, String property}) {
+  Binding _binding({required ModelBinding parentBinding,required PField schema,required String property}) {
     switch (schema.runtimeType) {
       case PString:
         return parentBinding.stringBinding(property: property);
@@ -32,7 +40,7 @@ mixin ConnectorBuilder {
     }
   }
 
-  ModelViewConverter _converter({PField schema, Type viewDataType}) {
+  ModelViewConverter _converter({required PField schema,required Type viewDataType}) {
     switch (schema.runtimeType) {
       case PInteger:
         return _intConverter(viewDataType);
